some bug modification

+++ src/Controllers/TaskMaterialController.php
@@ -67,7 +67,7 @@ class TaskMaterialController extends Controller
             return redirect('tasks.create')
                 ->withErrors($validator)
                 ->withInput();
-        } else {
+        } elseif($request->material_id) {
             // store
 
             foreach ($request->material_id as $key => $row) {
@@ -134,18 +134,19 @@ class TaskMaterialController extends Controller
                 'message' => TaskHelper::getStatusMessage('task_approver_edited')
             ]);
         }
-
-        $t = TaskMaterial::where('task_id', $request->task_id);
-        $t->delete();
-        foreach ($request->material_id as $key => $row) {
-            $attributes = [
-                'task_id' => $request->task_id,
-                'material_id' => $request->material_id[$key],
-                'material_qty' => $request->material_qty[$key],
-                'material_amount' => $request->material_amount[$key],
-                'material_note' => $request->material_note[$key],
-            ];
-            $taskmaterial = $this->taskmaterial->create($attributes);
+        if($request->material_id){
+            $t = TaskMaterial::where('task_id', $request->task_id);
+            $t->delete();
+            foreach ($request->material_id as $key => $row) {
+                $attributes = [
+                    'task_id' => $request->task_id,
+                    'material_id' => $request->material_id[$key],
+                    'material_qty' => $request->material_qty[$key],
+                    'material_amount' => $request->material_amount[$key],
+                    'material_note' => $request->material_note[$key],
+                ];
+                $taskmaterial = $this->taskmaterial->create($attributes);
+            }
         }
         //dd($request->all());
         try {
